+++ R/np.reghat.R
@@ -1844,7 +1844,7 @@ npreghat.formula <-
     tmf[[1]] <- as.name("model.frame")
     tmf[["formula"]] <- tt
     if (!is.null(data))
-      tmf[["data"]] <- substitute(data)
+      tmf[["data"]] <- data
     mf.args <- as.list(tmf)[-1L]
     mf <- do.call(stats::model.frame, mf.args, envir = environment(tt))